Created orbital analyzer package and shifted revisit, coverage and LOS code
